add Scala.js support
 - add cross-project Scala.js support to build
 - separate JVM-only code into JVM project
 - add BeforeAndAfterAll to HttpTest so WireMockServer will stop after tests complete
 - update some metadata in build definition
