+++ .travis.yml
@@ -4,7 +4,7 @@ jdk:
   - oraclejdk8
 
 scala:
-  - 2.12.4
+  - 2.12.5
 
 notifications:
   email:
