+++ jvm/src/main/scala/com/github/mideo/media/m3u8/io/FileSystem.scala
