+++ jvm/src/test/scala/StreamPlaylistTest.scala
@@ -1,12 +1,9 @@
 
 import com.github.mideo.M3U8ParserSuite
 import com.github.mideo.media.m3u8._
-import com.github.mideo.media.m3u8._
 import com.github.mideo.media.m3u8.domain._
-import org.scalatest.{Assertion, Failed, Outcome, Succeeded}
 
 import scala.io.Source
-import scala.concurrent.Future
 
 
 class StreamPlaylistTest extends M3U8ParserSuite {
