+++ jvm/src/test/scala/StreamPlaylistWriteTest.scala
