+++ jvm/src/test/scala/com/github/mideo/M3U8ParserSuite.scala
