+++ jvm/src/test/scala/com/github/mideo/media/m3u8/StreamPlaylistSectionTest.scala
