+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 1.0.4
+sbt.version = 1.1.2
