+++ shared/src/main/scala/com/github/mideo/media/m3u8/domain/Deserializers.scala
