--- src/main/scala/com/github/mideo/media/m3u8/package.scala
@@ -1,10 +1,0 @@
-package com.github.mideo.media
-
-import com.github.mideo.media.m3u8.domain.MediaStreamType
-
-package object m3u8 {
-  private[media] trait StreamPlaylist {
-    val mediaStreamType: Option[MediaStreamType]
-    override def toString: String
-  }
-}
